refacto: move ImageryLayers in its own module
